+++ uvicorn/main.py
@@ -406,7 +406,7 @@ def main(
     app_dir: str,
     h11_max_incomplete_event_size: int | None,
     factory: bool,
-    shutdown_delay: int = 0
+    shutdown_delay: float = 0
 ) -> None:
     run(
         app,
@@ -509,7 +509,7 @@ def run(
     app_dir: str | None = None,
     factory: bool = False,
     h11_max_incomplete_event_size: int | None = None,
-    shutdown_delay: int = 0
+    shutdown_delay: float = 0
 ) -> None:
     if app_dir is not None:
         sys.path.insert(0, app_dir)
